@@ -13,40 +13,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.1.2",
-    "@angular/common": "7.1.2",
-    "@angular/compiler": "7.1.2",
-    "@angular/core": "7.1.2",
-    "@angular/forms": "7.1.2",
-    "@angular/http": "7.1.2",
-    "@angular/platform-browser": "7.1.2",
-    "@angular/platform-browser-dynamic": "7.1.2",
-    "@angular/router": "7.1.2",
+    "@angular/animations": "7.2.12",
+    "@angular/common": "7.2.12",
+    "@angular/compiler": "7.2.12",
+    "@angular/core": "7.2.12",
+    "@angular/forms": "7.2.12",
+    "@angular/http": "7.2.12",
+    "@angular/platform-browser": "7.2.12",
+    "@angular/platform-browser-dynamic": "7.2.12",
+    "@angular/router": "7.2.12",
     "core-js": "2.6.0",
     "ngx-quill": "3.3.0",
-    "rxjs": "6.3.3",
-    "zone.js": "0.8.26"
+    "rxjs": "6.4.0",
+    "zone.js": "0.9.0"
   },
   "devDependencies": {
-    "@angular/cli": "7.1.2",
-    "@angular/compiler-cli": "7.1.2",
-    "@angular/language-service": "7.1.2",
-    "@types/jasmine": "3.3.0",
+    "@angular/cli": "7.3.8",
+    "@angular/compiler-cli": "7.2.12",
+    "@angular/language-service": "7.2.12",
+    "@types/jasmine": "3.3.12",
     "@types/jasminewd2": "2.0.6",
-    "@types/node": "10.12.12",
-    "codelyzer": "4.5.0",
-    "jasmine-core": "3.3.0",
+    "@types/node": "11.13.0",
+    "codelyzer": "5.0.0",
+    "jasmine-core": "3.4.0",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "3.1.3",
+    "karma": "4.0.1",
     "karma-chrome-launcher": "2.2.0",
-    "karma-cli": " 1.0.1",
-    "karma-coverage-istanbul-reporter": "2.0.4",
+    "karma-cli": "2.0.0",
+    "karma-coverage-istanbul-reporter": "2.0.5",
     "karma-jasmine": "2.0.1",
     "karma-jasmine-html-reporter": "1.4.0",
-    "protractor": "5.4.1",
-    "ts-node": "7.0.1",
-    "tslint": "5.11.0",
+    "protractor": "5.4.2",
+    "ts-node": "8.0.3",
+    "tslint": "5.15.0",
     "typescript": "3.1.2",
-    "@angular-devkit/build-angular": "0.11.2"
+    "@angular-devkit/build-angular": "0.13.8"
   }
 }
